pyproject.toml
[project]
name = "uncertaintylm"
version = "0.1.0"
description = ""
authors = [
    {name = "CatRespect",email = "alexej.feofanov@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "lm-polygraph (>=0.4.0,<0.5.0)",
    "transformers (>=4.53.3,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "huggingface-hub (>=0.34.1,<0.35.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "hf-xet (>=1.1.5,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
